HADOOP-16830. Reinstate the deleted FunctionsRaisingIOE class.
This is to allow hadoop-aws to compile without needing to move to the
moved interfaces; it will also allow anything external which used those
methods (unlikely) to keep working.

Also: throw synchronized at methods to get findbugs to STFU.

Tempting just to turn it off on the basis that it is overkill, but it could
maybe be correct.

Making all of MeanStatistics synchronized is irritating, as it will hurt
performance on what should be a lightweight class. But it is needed to
ensure samples and sum are consistently set.

Change-Id: I4c3e2726e1e97d705c55dbb43a507ea4d0e81e28